+++ weblate/trans/tests/test_machine.py
@@ -37,7 +37,9 @@ from weblate.trans.machine.apertium import (
 )
 from weblate.trans.machine.tmserver import AmagamaTranslation
 from weblate.trans.machine.microsoft import (
-    MicrosoftTranslation, MicrosoftCognitiveTranslation,
+    MicrosoftTranslation,
+    MicrosoftCognitiveTranslation,
+    MicrosoftTerminologyService
 )
 from weblate.trans.machine.google import GoogleTranslation, GOOGLE_API_ROOT
 from weblate.trans.machine.yandex import YandexTranslation
@@ -116,6 +118,111 @@ SAPTRANSLATIONHUB_JSON = '''
         }
     ]
 }
+'''.encode('utf-8')
+
+TERMINOLOGY_LANGUAGES = '''
+<s:Envelope xmlns:s="http://schemas.xmlsoap.org/soap/envelope/">
+  <s:Body>
+    <GetLanguagesResponse xmlns="http://api.terminology.microsoft.com/terminology">
+      <GetLanguagesResult xmlns:i="http://www.w3.org/2001/XMLSchema-instance">
+        <Language>
+          <Code>af-za</Code>
+        </Language>
+        <Language>
+          <Code>am-et</Code>
+        </Language>
+        <Language>
+          <Code>ar-dz</Code>
+        </Language>
+        <Language>
+          <Code>ar-eg</Code>
+        </Language>
+        <Language>
+          <Code>ar-sa</Code>
+        </Language>
+        <Language>
+          <Code>as-in</Code>
+        </Language>
+        <Language>
+          <Code>az-latn-az</Code>
+        </Language>
+        <Language>
+          <Code>be-by</Code>
+        </Language>
+        <Language>
+          <Code>bg-bg</Code>
+        </Language>
+        <Language>
+          <Code>bn-bd</Code>
+        </Language>
+        <Language>
+          <Code>bn-in</Code>
+        </Language>
+        <Language>
+          <Code>bs-cyrl-ba</Code>
+        </Language>
+        <Language>
+          <Code>bs-latn-ba</Code>
+        </Language>
+        <Language>
+          <Code>ca-es</Code>
+        </Language>
+        <Language>
+          <Code>ca-es-valencia</Code>
+        </Language>
+        <Language>
+          <Code>chr-cher-us</Code>
+        </Language>
+        <Language>
+          <Code>cs-cz</Code>
+        </Language>
+        <Language>
+          <Code>en-us</Code>
+        </Language>
+      </GetLanguagesResult>
+    </GetLanguagesResponse>
+  </s:Body>
+</s:Envelope>
+'''.encode('utf-8')
+TERMINOLOGY_TRANSLATE = '''
+<s:Envelope xmlns:s="http://schemas.xmlsoap.org/soap/envelope/">
+  <s:Body>
+    <GetTranslationsResponse xmlns="http://api.terminology.microsoft.com/terminology">
+      <GetTranslationsResult xmlns:i="http://www.w3.org/2001/XMLSchema-instance">
+        <Match>
+          <ConfidenceLevel>100</ConfidenceLevel>
+          <Count>8</Count>
+          <Definition i:nil="true"/>
+          <OriginalText>Hello World</OriginalText>
+          <Product i:nil="true"/>
+          <ProductVersion i:nil="true"/>
+          <Source i:nil="true"/>
+          <Translations>
+            <Translation>
+              <Language>cs-cz</Language>
+              <TranslatedText>Hello World</TranslatedText>
+            </Translation>
+          </Translations>
+        </Match>
+        <Match>
+          <ConfidenceLevel>100</ConfidenceLevel>
+          <Count>1</Count>
+          <Definition i:nil="true"/>
+          <OriginalText>Hello world.</OriginalText>
+          <Product i:nil="true"/>
+          <ProductVersion i:nil="true"/>
+          <Source i:nil="true"/>
+          <Translations>
+            <Translation>
+              <Language>cs-cz</Language>
+              <TranslatedText>Ahoj sv&#x11B;te.</TranslatedText>
+            </Translation>
+          </Translations>
+        </Match>
+      </GetTranslationsResult>
+    </GetTranslationsResponse>
+  </s:Body>
+</s:Envelope>
 '''.encode('utf-8')
 
 class MachineTranslationTest(TestCase):
@@ -268,6 +375,42 @@ class MachineTranslationTest(TestCase):
 
         machine = MicrosoftCognitiveTranslation()
         self.assert_translate(machine)
+
+    @httpretty.activate
+    def test_microsoft_terminology(self):
+        def request_callback(request, uri, headers):
+            if b'GetLanguages' in request.body:
+                return (200, headers, TERMINOLOGY_LANGUAGES)
+            return (200, headers, TERMINOLOGY_TRANSLATE)
+
+        cache.delete(
+            '{0}-languages'.format(MicrosoftTerminologyService().mtid)
+        )
+        httpretty.register_uri(
+            httpretty.POST,
+            'http://api.terminology.microsoft.com/Terminology.svc',
+            body=request_callback,
+            content_type='text/xml',
+        )
+        machine = MicrosoftTerminologyService()
+        self.assert_translate(machine)
+        self.assert_translate(machine, lang='cs_CZ')
+
+    @httpretty.activate
+    def test_microsoft_terminology_error(self):
+        cache.delete(
+            '{0}-languages'.format(MicrosoftTerminologyService().mtid)
+        )
+        httpretty.register_uri(
+            httpretty.POST,
+            'http://api.terminology.microsoft.com/Terminology.svc',
+            body='',
+            content_type='text/xml',
+            status=500,
+        )
+        machine = MicrosoftTerminologyService()
+        self.assertEqual(machine.supported_languages, [])
+        self.assert_translate(machine, empty=True)
 
     @override_settings(MT_GOOGLE_KEY='KEY')
     @httpretty.activate